edit: Move into filesystem/misc/
