+++ extractor/filesystem/misc/chrome/extensions/extensions.go
