+++ extractor/filesystem/misc/chrome/extensions/metadata.go
