+++ extractor/filesystem/misc/chrome/extensions/testdata/aapbdbdomjkkjkaonfhkkikfgjllcleb/2.0.16_0/manifest.json
