+++ extractor/filesystem/misc/chrome/extensions/testdata/ghbmnnjooekpmoecnnnilnnbdlolhkhi/1.89.1/manifest.json
