+++ extractor/filesystem/misc/chrome/extensions/testdata/invalid-manifest.json
