+++ extractor/filesystem/misc/chrome/extensions/testdata/nolocaleoekpmoecnnnilnnbdlolhkhi/1.89.1/manifest.json
